Unpack tgz files into pod's root.
https://github.com/CocoaPods/CocoaPods/issues/727

Zip files are usually packaged with their contents in the archive file's root.
Tgz files are usually packaged with only a folder in the root.
Make them both behave the same by moving that single folder's contents up.
This is basically the opposite of what Finder does (which creates a folder for the zip file's
contents, so you always unarchive into a folder).

Signed-off-by: Eloy Durán <eloy.de.enige@gmail.com>