Use curl's --netrc-optional option instead of --netrc
Using --netrc forces the user to provide authentication credentials through a
.netrc file in the user's home folder if the remote server requires
authentication.

However a user may want to provide the credentials right in the url which has
the advantage of not being dependent on having build machines set up with the
.netrc file. In an enterprise environment - potentially using many build
machines and having many team members - it can become a hassle to ensure proper
configuration of all build environments.

Ideally we should be able to specify the path to the .netrc file using the
--netrc-file option but that's for another time.

gh-55