Explicitly checkout submodules when needed
After doing some research, it looks like git 2.9 made a change so that command
line flags are now sent to the submodule commands as well, where pre-2.9
releases ignored those flags for submodules.

That seems like a great change on the surface. But unfortunately for us, it
means that using `--depth=1` and `--recursive` in the same command will almost
always result in a failed clone. This is because the submodule is also
initially checked out using `--depth=1` and so when it then tries to check out
the commit specified by the .gitmodules file, it fails because it doesn't see
that SHA in the git history. The only way it will succeed is if the specified
commit also happens to be the HEAD of master.

This issue can be fixed by removing the `--recursive` flag from the initial
git command, and then running `git submodules --init --recursive` inside the
repo to pull down the submodules.