Validate the ref that was checked out was a tag
Previously if you had a repo with a branch and a tag with the same name,
using `git clone --branch` would prefer the branch, which may or may not
be the same source as the tag. Now we checkout the tag unambiguously.
The biggest downside here is we no longer do a shallow clone in this
case.